Configure stb_image to use an arena
Probably a bad idea, because arena here wastes quite a bunch of memory
without giving any benefits back. (The cost of calling malloc here is
way too low in comparison to the useful work getting done.)

I just wanted to see, if it would be possible to come up with an
interface that does not depend on the specific arena struct type. You
might want this in case you have multiple independent libraries and each
of them uses its own type for the arena, but within the calling code you
want to use a single arena for all of them.

Just using a "void *" as the arena type and accessing the arena data at
pre-agreed offsets might be a better solution?