Try to estimate stb_image memory usage
Turned out to be a dumb idea. Just width and height of the input image
alone give us a very loose upper estimate of how much memory stb_image
is going to use. For a JPEG we don't know upfront which kind of chroma
subsapling it uses. For a PNG we don't know if it has an alpha channel
and whether it uses a palette. Not even mentioning that such estimate is
very brittle and only works for this specific version of stb_image.