Fix error clamping in Floyd-Steinberg diffusion
Calcualte the max and min error based on the max distance between two
successive colors from the palette. Store the un-clamped errors instead
of pixel values with accumulated error added. Only clamp an error value
when adding it to the current pixel value.

Additionally, don't store error values for the whole image. Only store
errors for the current and the next pixel rows.